backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.10
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
boto3==1.34.14
celery[redis]==5.3.4
redis==4.6.0
sendgrid==6.10.0
stripe==7.10.0
sentry-sdk[fastapi]==1.39.1
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
shortuuid==1.0.11
python-magic==0.4.27
httpx==0.26.0
slowapi==0.1.9
bleach==6.1.0
Pillow==10.2.0
PyPDF2==3.0.1
pdf2image==1.17.0
pyotp==2.9.0
qrcode[pil]==7.4.2
pandas==2.1.4
openpyxl==3.1.2
aiofiles==23.2.1
aiosqlite==0.19.0
gunicorn==21.2.0